Implemented added/update/removal logic into repository

+++ JMBlogWebApp/repository/image_db_repository.py
@@ -6,11 +6,14 @@ class ImageDbRepository(ImageRepository):
     def __init__(self):
         pass
 
-    def add_image(self, added_image):
-        added_image.save(os.path.join('static/images', added_image.filename))
+    def add_image(self, blog_post, added_image):
+        if added_image.filename != '':
+            added_image.save(os.path.join('static/images', added_image.filename))
+            blog_post.img_path = 'images/{}'.format(added_image.filename)
 
     def remove_image(self, blog_post):
-        if os.path.exists(os.path.join('static', blog_post.img_path)):
+        if (blog_post.img_path is not None and
+            os.path.exists(os.path.join('static', blog_post.img_path))):
             os.remove(os.path.join('static', blog_post.img_path))
         blog_post.img_path = None
